refactor: 도메인별 가격을 포장한 객체 정의 및 테스트 생성

+++ src/main/java/nextstep/subway/auth/domain/Money.java
@@ -49,7 +49,7 @@ public class Money {
         return Money.from(this.amount - (this.amount * rate));
     }
 
-    public Money mul(double target) {
+    public Money mul(int target) {
         return Money.from(this.amount * target);
     }
