+++ src/test/java/nextstep/subway/auth/domain/MoneyTest.java
@@ -61,7 +61,7 @@ class MoneyTest {
     @Test
     void multiply_money_success() {
         final double source = 100;
-        final double target = 2;
+        final int target = 2;
         final double expected = 200;
         final Money sourceMoney = Money.from(source);
         assertThat(sourceMoney.mul(target)).isEqualTo(Money.from(expected));
